Split arithmetic ops into subclasses (core & numpy)
