+++ temporian/core/operators/all_operators.py
@@ -3,7 +3,7 @@
 # pylint: disable=unused-import
 from temporian.core.operators.arithmetic import divide
 from temporian.core.operators.arithmetic import multiply
-from temporian.core.operators.arithmetic import substract
+from temporian.core.operators.arithmetic import subtract
 from temporian.core.operators.arithmetic import sum
 
 from temporian.core.operators.glue import glue
